prepare project:

+++ database/migrations/2024_11_27_000001_create_users_table.php
@@ -13,15 +13,17 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name', 100);
-            $table->string('email', 150)->unique();
-            $table->string('password');
-            $table->string('phone', 20)->nullable();
-            $table->foreignId('role_id')->nullable()->constrained('roles')->onDelete('set null');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('name', 100);
+                $table->string('email', 150)->unique();
+                $table->string('password');
+                $table->string('phone', 20)->nullable();
+                $table->foreignId('role_id')->nullable()->constrained('roles')->onDelete('set null');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
